// food-recipe-react-native/app/constants/types/screen.ts
export enum SCREEN {
  RECIPE_LIST = 'RECIPE_LIST',
  RECIPE_SEARCH = 'RECIPE_SEARCH',
  RECIPE_DETAIL = 'RECIPE_DETAIL',
  RECIPE_PRICE = 'RECIPE_PRICE',
  RECIPE_NUTRITION = 'RECIPE_NUTRITION',
  VIEW_IMAGE = 'VIEW_IMAGE',
  INGREDIENT_BENEFIT = 'INGREDIENT_BENEFIT',
  INGREDIENT_WARNING = 'INGREDIENT_WARNING',
  TIP = 'TIP'
}